atoi base

+++ srcs/ft_uitoa.c
@@ -6,7 +6,7 @@
 /*   By: jacens <jacens@student.le-101.fr>          +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/10/10 16:16:12 by jacens       #+#   ##    ##    #+#       */
-/*   Updated: 2019/11/18 13:37:50 by jacens      ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/11/22 13:46:36 by jacens      ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -52,7 +52,7 @@ char			*ft_uitoa(unsigned int n)
 	unsigned long long		i;
 	long long				j;
 
-	nb = n;
+	nb = (unsigned long long)n;
 	j = -1;
 	if (!(ptr = malloc(ft_ucount_nbr(nb) + 1)))
 		return (NULL);
